requirements/base.txt
Django==1.10.5
django-allauth==0.30.0
psycopg2==2.6.2
django-model-utils==2.6.1
Pillow==4.0.0
qrcode==5.3
channels==1.0.2
djangorestframework==3.5.3
